load.py uses API instead of DB directly
This was done in preparation of separating the backend from the ETL
process itself. As a consequence, a lot of the endpoints now require
id:s, which isn't really ideal. Should change the db so that the PKs /
IDs are not the ones provided in the data, but instead generated by the
backend itself. Also, load.py utilizes POSTs, while they should probably
be PUTs. Also, frontend can't do all that much, now.