separated schemas into files
This is useful not only because they are easier to work with when in
individual files, but also because we now import through __init__.py.
This way our IDE doesn't show us baseclasses and other stuff that isn't
relevant in other parts of the application, unless we explicitly want
to. (Granted, the same could be achieved with __all__, too.)